kill tournamentFunctions, more function organization
