+++ DiscordBot/functions/confirmCommand.ts
@@ -1,0 +1,43 @@
+import { randomUUID } from "crypto";
+import { ActionRowBuilder, ButtonBuilder, ButtonStyle, ChatInputCommandInteraction, Message, MessageComponentInteraction } from "discord.js";
+
+export default async function confirmCommand (m: Message | ChatInputCommandInteraction, content: string): Promise<boolean> {
+    const ids = {
+        yes: randomUUID(),
+        no: randomUUID(),
+    }
+    const message = await m.channel!.send({
+        content,
+        components: [
+            new ActionRowBuilder<ButtonBuilder>()
+                .addComponents(
+                    new ButtonBuilder()
+                        .setCustomId(ids.yes)
+                        .setLabel("Yes")
+                        .setStyle(ButtonStyle.Success),
+                    new ButtonBuilder()
+                        .setCustomId(ids.no)
+                        .setLabel("No")
+                        .setStyle(ButtonStyle.Danger)),
+        ],
+    });
+
+    return new Promise<boolean>(resolve => {
+        const filter = (i: MessageComponentInteraction) => i.user.id === (m instanceof Message ? m.author.id : m.user.id);
+        const confirmationCollector = m.channel!.createMessageComponentCollector({ filter, time: 6000000 });
+        confirmationCollector.on("collect", async (msg: Message | MessageComponentInteraction) => {
+            if (msg instanceof MessageComponentInteraction) {
+                if (msg.customId === ids.yes) {
+                    await message.delete();
+                    confirmationCollector.stop();
+                    resolve(true);
+                }
+                else if (msg.customId === ids.no) {
+                    await message.delete();
+                    confirmationCollector.stop();
+                    resolve(false);
+                }
+            }
+        });
+    });
+}
