+++ Models/tournaments/mappools/mappool.ts
@@ -47,37 +47,4 @@ export class Mappool extends BaseEntity {
     @OneToMany(() => MappoolSlot, mappoolSlot => mappoolSlot.mappool)
     slots!: MappoolSlot[];
 
-    static search (tournament: Tournament, poolText: string = "", getStageRound: boolean = false, getSlots: boolean = false, getMaps: boolean = false) {
-        const q = this.createQueryBuilder("mappool")
-        if (getStageRound) {
-            q.leftJoinAndSelect("mappool.stage", "stage")
-            q.leftJoinAndSelect("mappool.round", "round")
-        } else
-            q.leftJoin("mappool.stage", "stage")
-
-        if (getSlots) {
-            q.leftJoinAndSelect("mappool.slots", "slot")
-            q.leftJoinAndSelect("slot.maps", "mappoolMap")
-            if (getMaps) {
-                q.leftJoinAndSelect("mappoolMap.customBeatmap", "customBeatmap")
-                q.leftJoinAndSelect("mappoolMap.customMappers", "customMapper")
-                q.leftJoinAndSelect("mappoolMap.testplayers", "testplayer")
-                q.leftJoinAndSelect("mappoolMap.jobPost", "jobPost")
-                q.leftJoinAndSelect("mappoolMap.beatmap", "beatmap")
-                q.leftJoinAndSelect("beatmap.beatmapset", "beatmapset")
-            }
-        }
-        return q
-            .where("stage.tournament = :tournament")
-            .andWhere(new Brackets(qb => {
-                qb.where("mappool.name LIKE :criteria")
-                    .orWhere("mappool.abbreviation LIKE :criteria");
-            }))
-            .setParameters({
-                tournament: tournament.ID,
-                criteria: `%${poolText}%`,
-            })
-            .getMany();
-    }
-
 }
